restructure and rename response
