+++ crates/bitwarden/src/secrets_manager/client_secrets.rs
@@ -5,7 +5,7 @@ use crate::{
         list_secrets_by_project, update_secret, SecretCreateRequest, SecretGetRequest,
         SecretIdentifiersByProjectRequest, SecretIdentifiersRequest, SecretIdentifiersResponse,
         SecretPutRequest, SecretResponse, SecretsDeleteRequest, SecretsDeleteResponse,
-        SecretsGetByIdsResponse, SecretsGetRequest,
+        SecretsGetRequest, SecretsResponse,
     },
     Client,
 };
@@ -19,10 +19,7 @@ impl<'a> ClientSecrets<'a> {
         get_secret(self.client, input).await
     }
 
-    pub async fn get_by_ids(
-        &mut self,
-        input: SecretsGetRequest,
-    ) -> Result<SecretsGetByIdsResponse> {
+    pub async fn get_by_ids(&mut self, input: SecretsGetRequest) -> Result<SecretsResponse> {
         get_secrets_by_ids(self.client, input).await
     }
