+++ crates/bitwarden/src/secrets_manager/secrets/mod.rs
@@ -13,11 +13,11 @@ pub use delete::{SecretsDeleteRequest, SecretsDeleteResponse};
 pub(crate) use get::get_secret;
 pub use get::SecretGetRequest;
 pub(crate) use get_by_ids::get_secrets_by_ids;
-pub use get_by_ids::{SecretsGetByIdsResponse, SecretsGetRequest};
+pub use get_by_ids::SecretsGetRequest;
 pub(crate) use list::{list_secrets, list_secrets_by_project};
 pub use list::{
     SecretIdentifiersByProjectRequest, SecretIdentifiersRequest, SecretIdentifiersResponse,
 };
-pub use secret_response::SecretResponse;
+pub use secret_response::{SecretResponse, SecretsResponse};
 pub(crate) use update::update_secret;
 pub use update::SecretPutRequest;
