+++ crates/sdk-schemas/src/main.rs
@@ -101,7 +101,7 @@ fn main() -> Result<()> {
         bitwarden::auth::response::PasswordLoginResponse,
         bitwarden::secrets_manager::secrets::SecretIdentifiersResponse,
         bitwarden::secrets_manager::secrets::SecretResponse,
-        bitwarden::secrets_manager::secrets::SecretsGetByIdsResponse,
+        bitwarden::secrets_manager::secrets::SecretsResponse,
         bitwarden::secrets_manager::secrets::SecretsDeleteResponse,
     };
